[SCOPE CREEP] Implemented ability to lookup keys with periods.
This is definitely scope creep to actually knock up support for looking a key
with a period in it!! Ideally, I should have just left it as a
query/observation, but had the itch to highlight and _"just do it!"_ (Can
either be appreciated or seen as a warning sign).

Change allows the escaped Full Stop to pass through as part of the Key lookup,
instead of splitting and having a false negative look up.

Issue: #7.